Cleanup openshift-route-controller-manager
