+++ control-plane-operator/controllers/hostedcontrolplane/v2/assets/openshift-route-controller-manager/config.yaml
@@ -58,17 +58,9 @@ data:
       bindAddress: 0.0.0.0:8443
       bindNetwork: ""
       certFile: /etc/kubernetes/certs/tls.crt
-      cipherSuites:
-      - TLS_ECDHE_ECDSA_WITH_AES_128_GCM_SHA256
-      - TLS_ECDHE_RSA_WITH_AES_128_GCM_SHA256
-      - TLS_ECDHE_ECDSA_WITH_AES_256_GCM_SHA384
-      - TLS_ECDHE_RSA_WITH_AES_256_GCM_SHA384
-      - TLS_ECDHE_ECDSA_WITH_CHACHA20_POLY1305_SHA256
-      - TLS_ECDHE_RSA_WITH_CHACHA20_POLY1305_SHA256
       clientCA: /etc/kubernetes/client-ca/ca.crt
       keyFile: /etc/kubernetes/certs/tls.key
       maxRequestsInFlight: 0
-      minTLSVersion: VersionTLS12
       requestTimeoutSeconds: 0
 kind: ConfigMap
 metadata:
